# api/openapi/atlas.openapi.json
{
  "components": {
    "schemas": {
...
      "DomainDetailResponse": {
        "description": "Domain detail resource.",
        "properties": {
          "issue_areas": {
            "items": {
              "$ref": "#/components/schemas/IssueAreaResponse"
            },
            "title": "Issue Areas",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "DomainDetailResponse",
        "type": "object"
      },
      "DomainListResponse": {
        "description": "Collection response for domains.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DomainResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "DomainListResponse",
        "type": "object"
      },
      "DomainResponse": {
        "description": "Domain summary resource.",
        "properties": {
          "issue_area_count": {
            "title": "Issue Area Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "issue_area_count"
        ],
        "title": "DomainResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IssueAreaListResponse": {
        "description": "Collection response for issue areas.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueAreaResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "IssueAreaListResponse",
        "type": "object"
      },
      "IssueAreaResponse": {
        "description": "Canonical issue area resource.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "domain"
        ],
        "title": "IssueAreaResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "atlas@example.com",
      "name": "Rebuilding America",
      "url": "https://atlas.example.com/"
    },
    "description": "Atlas exposes place-first civic research data through a typed REST API. The public contract includes entities, sources, issue areas, domains, place profiles, issue signals, discovery runs, and anonymous data flags.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Structured civic research data for Atlas places, entities, sources, and discovery runs.",
    "title": "Atlas REST API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/domains": {
      "get": {
        "description": "List top-level Atlas issue domains.",
        "operationId": "listDomains",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainListResponse"
                }
              }
            },
            "description": "A paginated collection of Atlas domains."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List domains",
        "tags": [
          "domains"
        ]
      }
    },
    "/api/domains/{domain_slug}": {
      "get": {
        "description": "Return a single Atlas domain and its issue areas.",
        "operationId": "getDomain",
        "parameters": [
          {
            "in": "path",
            "name": "domain_slug",
            "required": true,
            "schema": {
              "title": "Domain Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainDetailResponse"
                }
              }
            },
            "description": "The requested Atlas domain."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a domain",
        "tags": [
          "domains"
        ]
      }
    },
...
    "/api/issue-areas": {
      "get": {
        "description": "List Atlas issue areas or resolve natural-language issue queries against the taxonomy.",
        "operationId": "listIssueAreas",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueAreaListResponse"
                }
              }
            },
            "description": "A paginated collection of Atlas issue areas."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List issue areas",
        "tags": [
          "issue-areas"
        ]
      }
    },
    "/api/issue-areas/{issue_area_slug}": {
      "get": {
        "description": "Return one Atlas issue area by slug.",
        "operationId": "getIssueArea",
        "parameters": [
          {
            "in": "path",
            "name": "issue_area_slug",
            "required": true,
            "schema": {
              "title": "Issue Area Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueAreaResponse"
                }
              }
            },
            "description": "The requested Atlas issue area."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get an issue area",
        "tags": [
          "issue-areas"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://atlas.example.com"
    },
    {
      "url": "http://localhost:8000"
    },
    {
      "url": "/"
    }
  ],
...
}